requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.11.2
black==23.1.0
blis==0.7.11
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.3
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
confection==0.1.4
constantly==23.10.4
cryptography==42.0.4
cssselect==1.2.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
dataclasses-json==0.6.4
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
docopt==0.6.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
executing==1.2.0
fastapi==0.103.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2024.2.0
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.0
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.1.1
incremental==22.10.0
ipykernel==6.24.0
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
Jinja2==3.1.3
jmespath==1.0.1
json-parser==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.0
jupyter_core==5.3.1
kubernetes==29.0.0
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.25
langchain-openai==0.0.6
langchainhub==0.1.14
langcodes==3.3.0
langgraph==0.0.25
langsmith==0.1.5
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib-inline==0.1.6
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.2.1
numpy==1.24.2
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
parsel==1.8.1
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
pipreqs==0.4.11
platformdirs==3.0.0
posthog==3.4.2
preshed==3.0.9
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.25.3
psutil==5.9.5
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyDispatcher==2.0.7
Pygments==2.15.1
pyOpenSSL==24.0.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
queuelib==1.6.2
regex==2023.12.25
requests==2.28.2
requests-file==2.0.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.7.4
spacy-alignments==0.9.1
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
spacy-transformers==1.3.4
SQLAlchemy==2.0.21
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.6.0
tldextract==5.1.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.36.2
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typer==0.9.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.23.2
w3lib==2.1.2
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.6
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.8.2
zipp==3.17.0
zope.interface==6.2
